requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Django==3.1.6
django-crispy-forms==1.11.0
django-livereload-server==0.3.2
djangorestframework==3.12.2
pytz==2021.1
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
tornado==6.1
